backend/requirements.txt
fastapi>=0.110
uvicorn[standard]>=0.27
websockets>=12.0
pydantic>=2.5
boto3>=1.34
strands-agents>=0.1
strands-agents-tools>=0.1
sqlalchemy>=2.0
psycopg2-binary>=2.9
opensearch-py>=2.4
requests-aws4auth>=1.2
graphrag-toolkit>=3.16
python-dotenv>=1.0
